requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2

# Econometric Analysis
statsmodels==0.14.0
linearmodels==5.3

# Feature Engineering
feature-engine==1.6.1
shap==0.43.0

# Imbalanced Learning
imbalanced-learn==0.11.0

# Hyperparameter Optimization
optuna==3.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.2
